backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
opencv-python-headless
Pillow
chromadb
pydantic
sqlalchemy
supabase
paramiko
supabase
boto3
